requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=6.0.0

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Azure services
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-cosmos>=4.5.0
openai>=1.12.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Speech (local STT)
faster-whisper>=1.0.0

# Speech (local TTS)
edge-tts>=6.1.0

